Move connector-specific CREATE,DROP TABLE tests to smoke tests
Most of the tests for CREATE are connector-specific and so they should
go to `AbstractTestIntegrationSmokeTest`.
