+++ presto-accumulo/src/test/java/com/facebook/presto/accumulo/TestAccumuloDistributedQueries.java
@@ -55,41 +55,6 @@ public class TestAccumuloDistributedQueries
     }
 
     @Override
-    public void testCreateTableAsSelect()
-    {
-        // This test is overridden due to Function "UUID" not found errors
-        // Some test cases from the base class are removed
-
-        assertUpdate("CREATE TABLE test_create_table_as_if_not_exists (a bigint, b double)");
-        assertTrue(getQueryRunner().tableExists(getSession(), "test_create_table_as_if_not_exists"));
-        assertTableColumnNames("test_create_table_as_if_not_exists", "a", "b");
-
-        assertUpdate("CREATE TABLE IF NOT EXISTS test_create_table_as_if_not_exists AS SELECT UUID() AS uuid, orderkey, discount FROM lineitem", 0);
-        assertTrue(getQueryRunner().tableExists(getSession(), "test_create_table_as_if_not_exists"));
-        assertTableColumnNames("test_create_table_as_if_not_exists", "a", "b");
-
-        assertUpdate("DROP TABLE test_create_table_as_if_not_exists");
-        assertFalse(getQueryRunner().tableExists(getSession(), "test_create_table_as_if_not_exists"));
-
-        this.assertCreateTableAsSelect(
-                "test_group",
-                "SELECT orderstatus, sum(totalprice) x FROM orders GROUP BY orderstatus",
-                "SELECT count(DISTINCT orderstatus) FROM orders");
-
-        this.assertCreateTableAsSelect(
-                "test_with_data",
-                "SELECT * FROM orders WITH DATA",
-                "SELECT * FROM orders",
-                "SELECT count(*) FROM orders");
-
-        this.assertCreateTableAsSelect(
-                "test_with_no_data",
-                "SELECT * FROM orders WITH NO DATA",
-                "SELECT * FROM orders LIMIT 0",
-                "SELECT 0");
-    }
-
-    @Override
     public void testDelete()
     {
         // Deletes are not supported by the connector
