+++ presto-cassandra/src/test/java/com/facebook/presto/cassandra/TestCassandraIntegrationSmokeTest.java
@@ -96,6 +96,12 @@ public class TestCassandraIntegrationSmokeTest
         return false;
     }
 
+    @Override
+    public void testCreateTable()
+    {
+        // this connector does not support creating tables (only CREATE TABLE AS SELECT is supported)
+    }
+
     @Test
     public void testPartitionKeyPredicate()
     {
