+++ presto-redis/src/test/java/com/facebook/presto/redis/TestRedisIntegrationSmokeTest.java
@@ -45,4 +45,22 @@ public class TestRedisIntegrationSmokeTest
     {
         embeddedRedis.close();
     }
+
+    @Override
+    public void testCreateTable()
+    {
+        // this connector does not support creating tables
+    }
+
+    @Override
+    public void testCreateTableAsSelect()
+    {
+        // this connector does not support creating tables
+    }
+
+    @Override
+    public void createTableWithUnicode()
+    {
+        // this connector does not support creating tables
+    }
 }
