+++ presto-thrift-connector/src/test/java/com/facebook/presto/connector/thrift/integration/TestThriftIntegrationSmokeTest.java
@@ -39,4 +39,22 @@ public class TestThriftIntegrationSmokeTest
                 .row("sf1");
         assertContains(actualSchemas, resultBuilder.build());
     }
+
+    @Override
+    public void testCreateTable()
+    {
+        // this connector does not support creating tables
+    }
+
+    @Override
+    public void testCreateTableAsSelect()
+    {
+        // this connector does not support creating tables
+    }
+
+    @Override
+    public void createTableWithUnicode()
+    {
+        // this connector does not support creating tables
+    }
 }
